+++ fiddles/three/fiddle-0007-Stars/src/metadata.js
@@ -1,0 +1,5 @@
+let metadata = {
+    urls: {
+        github: 'https://github.com/bradyhouse/house/tree/master/fiddles/three/fiddle-0007-Stars'
+    }
+};